+++ src/main/java/org/sopt/carena/healthreport/application/service/ExtractTextFromImageService.java
@@ -3,7 +3,7 @@ package org.sopt.carena.healthreport.application.service;
 import org.sopt.carena.healthreport.application.dto.command.ExtractTextCommand;
 import org.sopt.carena.healthreport.application.dto.view.ExtractedTextView;
 import org.sopt.carena.healthreport.application.port.in.ExtractTextFromImageUseCase;
-import org.sopt.carena.infrastructure.ocr.port.OcrPort;
+import org.sopt.carena.healthreport.application.port.out.OcrPort;
 import org.sopt.carena.healthreport.application.service.util.OcrHealthReportParser;
 import org.springframework.stereotype.Service;
