+++ src/main/java/org/sopt/carena/infrastructure/ocr/adapter/OcrAdapter.java
@@ -3,7 +3,7 @@ package org.sopt.carena.infrastructure.ocr.adapter;
 import org.sopt.carena.infrastructure.ocr.adapter.client.OcrClient;
 import org.sopt.carena.infrastructure.ocr.dto.OcrResponse;
 import org.sopt.carena.healthreport.application.dto.command.ExtractTextCommand;
-import org.sopt.carena.infrastructure.ocr.port.OcrPort;
+import org.sopt.carena.healthreport.application.port.out.OcrPort;
 import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
